QMAPS-2649 FloatingButton, time, global
