+++ @types/global.d.ts
@@ -18,6 +18,8 @@ declare global {
       ) => void;
       updateHash: (hash: string) => void;
     };
+
+    getLang: () => { locale: string; code: string; fallback?: string };
   }
 }
 export {};
